creating validation.js and including login validation.
